+++ docs/release-notes.md
@@ -25,6 +25,12 @@ of `zizmor`.
 
     Many thanks to @mostafa for implementing this improvement!
 
+### Bug Fixes 🐛
+
+* Fixed a bug where `zizmor` would fail instead of analyzing single-file
+  inputs that lacked an explicit parent path component, e.g.
+  `zizmor foo.yml` instead of `zizmor ./foo.yml` (#1212)
+
 ## 1.14.2
 
 ### Bug Fixes 🐛